# backend/fastapi/openapi-fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workma!te AI API",
    "description": "회의체 운영 AI Agent 서비스의 **AI · 실시간 · 지식그래프** 백엔드 (FastAPI).\n\n- **인증**: Spring이 발급한 JWT access token을 `Authorize`(HTTPBearer)에 입력하세요. (`JWT_SECRET`을 Spring과 공유, HS256)\n- **스트리밍**: 다수 엔드포인트가 SSE(`text/event-stream`)로 토큰 단위 응답합니다.\n- **데이터 스코프**: AI 도구는 사용자의 소속 회의체로 접근을 제한합니다(IDOR 이중 방어).\n- 인증·도메인 CRUD는 별도 Spring Boot 서버(`/api/v1`)가 담당합니다.",
    "contact": {
      "name": "Team No.9"
    },
    "license": {
      "name": "Internal"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/session/chat": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "세션 전용 챗봇 — SSE 스트리밍",
        "description": "세션 상태에 따라 분기하는 회의 전용 챗봇.\n\nrequire_view_by_session 가드로 세션 열람 권한을 확인하며, 응답은\nSSE(text/event-stream)로 스트리밍한다.\nscheduled → 기본 정보(일정·참석자·안건)만 안내\nongoing   → 진행 중 안내\nended     → 회의록 미생성 안내\narchived  → Neo4j RAG로 회의록 기반 답변",
        "operationId": "session_chat_api_agent_session_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/agent/supervisor/chat": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "슈퍼바이저 챗봇 — SSE 스트리밍",
        "description": "LLM이 의도를 분류해 하위 에이전트로 라우팅하는 통합 슈퍼바이저 챗봇.\n\n탈옥·유해 요청은 최전선에서 차단하고, 일일 토큰 예산을 검사한다. 멀티테넌트\n검색 스코프로 본인 소속 회의체만 조회하도록 강제(IDOR 방지)하며, 응답은\nSSE(text/event-stream)로 스트리밍한다.",
        "operationId": "supervisor_chat_api_agent_supervisor_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/chats/{context_type}/{context_id}": {
      "get": {
        "tags": [
          "chat_history"
        ],
        "summary": "대화 기록 조회",
        "description": "특정 컨텍스트의 대화 기록을 조회합니다(keyset 페이지네이션 지원).\n\n본인 스레드만 조회 — user_id == current_user.id로 스코프가 한정된다.",
        "operationId": "get_chat_history_api_chats__context_type___context_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Type"
            }
          },
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageOut"
                  },
                  "title": "Response Get Chat History Api Chats  Context Type   Context Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat_history"
        ],
        "summary": "대화 메시지 저장",
        "description": "사용자 메시지 1건을 본인 스레드에 저장합니다.\n\nuser 역할만 허용 — assistant/agent 메시지는 서버가 스트림 종료 시 직접 저장한다(H-5).",
        "operationId": "save_message_api_chats__context_type___context_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Type"
            }
          },
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat_history"
        ],
        "summary": "대화 기록 전체 삭제",
        "description": "특정 컨텍스트의 대화 기록을 전체 삭제합니다. 본인 스레드만(user_id 스코프) 삭제된다.",
        "operationId": "clear_chat_history_api_chats__context_type___context_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "context_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Type"
            }
          },
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload/chat": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "채팅 첨부파일 업로드",
        "description": "채팅 첨부파일을 R2에 업로드하고 chat_messages 테이블에 저장합니다.\n\nmeeting_id/session_id가 있으면 해당 회의체 조회 권한 필요. 파일 크기 상한 초과 시 413.",
        "operationId": "upload_chat_file_api_upload_chat_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_chat_file_api_upload_chat_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentChatRequest": {
        "properties": {
          "meeting_id": {
            "type": "integer",
            "title": "Meeting Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "chat_history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat History",
            "default": []
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "message"
        ],
        "title": "AgentChatRequest"
      },
...
      "Body_upload_chat_file_api_upload_chat_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type"
          },
          "meeting_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "thread_id"
        ],
        "title": "Body_upload_chat_file_api_upload_chat_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageIn": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "MessageIn"
      },
      "MessageOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}